+++ assets/media/class.js
@@ -5,19 +5,33 @@
 (function () {
 	const vscode = acquireVsCodeApi();
 
-	let ids = [
-		"startClass",
-		"arrowTo",
-		"noteLeft",
-		"noteRight"
-	]
+	let linkPropStatus = 0
+	let linkPropStatusId = "linkStructProperty"
+	let linkPropStatus0Str = "link struct property"
+	let linkPropStatus1Str = "property link to"
 
-	for (i = 0; i < ids.length; i++) {
-		document.getElementById(ids[i]).addEventListener('click', (e) => {
+	buttons = document.getElementsByClassName("add-color-button")
+	for (i = 0; i < buttons.length; i++) {
+		buttons[i].addEventListener('click', (e) => {
 			// console.log(e.srcElement)
 			// console.log(e.srcElement.innerText)
 			// console.log(e.srcElement.id)
-			vscode.postMessage({ type: 'functionSelected', value: e.srcElement.id });
+			if (e.srcElement.id != linkPropStatusId) {
+				document.getElementById(linkPropStatusId).innerText = linkPropStatus0Str
+				linkPropStatus = 0
+			}
+
+			console.log(e.srcElement.id)
+			if (e.srcElement.id == linkPropStatusId) {
+				if (e.srcElement.innerText == linkPropStatus0Str) {
+					e.srcElement.innerText = linkPropStatus1Str
+					linkPropStatus = 0
+				} else if (e.srcElement.innerText == linkPropStatus1Str) {
+					e.srcElement.innerText = linkPropStatus0Str
+					linkPropStatus = 1
+				}
+			}
+			vscode.postMessage({ type: 'functionSelected', value: e.srcElement.id, status: linkPropStatus });
 		});
 	}